+++ webui/src/modules/nodes/nodesPage/NodesContent/ReplicasetCard/instanceBlock/IpAddressLabel/IpAddressLabel.tsx
@@ -1,10 +1,9 @@
 import React from "react";
 import cn from "classnames";
 
-import { LinkSystemIcon } from "shared/icons/LinkSystemIcon";
 import { InfoNoData } from "shared/ui/InfoNoData/InfoNoData";
 import { useTranslation } from "shared/intl";
-import { HiddenWrapper } from "shared/ui/HiddenWrapper/HiddenWrapper";
+import { Clippable } from "shared/ui/Clippable/Clippable";
 
 import styles from "./IpAddressLabel.module.scss";
 
@@ -22,9 +21,8 @@ export const IpAddressLabel: React.FC<IpAddressLabelProps> = (props) => {
     return <InfoNoData text={translation.components.infoNoData.label} />;
 
   return (
-    <div className={cn(styles.container, className)}>
-      <LinkSystemIcon width={16} height={16} className={styles.icon} />
-      <HiddenWrapper>{address}</HiddenWrapper>
-    </div>
+    <Clippable className={cn(styles.container, className)} text={address}>
+      {address}
+    </Clippable>
   );
 };